[refactor](be) Move PackedUInt128 to packed_int128.h alongside PackedInt128
